Cmake and reorg
