+++ src/main.cpp
@@ -2,9 +2,15 @@
 #include <iostream>
 #include "UbuntuCloudFactory.hpp"
 
-typedef std::unique_ptr<UbuntuCloudInterface> UCI_Ptr;
-
-void printHelp() { }
+void printHelp() {
+  std::cout << "Usage: ubuntu-version-fetcher [OPTION]\n"
+            << "Options:\n"
+            << "  --list-releases     List all supported Ubuntu releases\n"
+            << "  --current-lts       Show the current Ubuntu LTS version\n"
+            << "  --sha256 RELEASE    Get the SHA256 hash of disk1.img for the specified release\n"
+            // << "  --url URL           Use custom Simplestreams URL\n"
+            << "  --help              Display this help and exit\n";
+}
 
 int main(int argc, char* argv[]) {
   std::ios_base::sync_with_stdio(false);
@@ -16,9 +22,9 @@ int main(int argc, char* argv[]) {
   curl_global_init(CURL_GLOBAL_DEFAULT);
   std::string option = argv[1];
 
-  UCI_Ptr fetcher = UbuntuCloudFactory::createFetcher();
+  std::unique_ptr<UbuntuCloudInterface> fetcher = UbuntuCloudFactory::createFetcher();
   if (fetcher == nullptr) {
-    std::cerr << "Failed to create cloud fetcher instance.\n";
+    std::cerr << "Failed to create cloud fetcher instance." << std::endl;
     return 1;
   }
   int return_code { 0 };
